+++ package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "@contentful/rich-text-html-renderer": "^15.13.1",
+    "@emailjs/browser": "^3.7.0",
     "@headlessui/react": "^1.6.6",
     "contentful": "^9.1.34",
     "gatsby": "^4.17.2",
