+++ src/pages/contact.js
@@ -1,11 +1,7 @@
 import React from "react";
+import ContactForm from "../components/ContactForm";
 
 export default function contact() {
-  const sendEmail = (e) => {
-    e.preventDefault();
-    alert("thanks for submitting the form");
-  };
-
   return (
     <>
       <div className="px-4 sm:px-6 d:px-12 lg:px-24 py-16 lg:py-24 flex flex-col justify-center items-center">
@@ -23,54 +19,7 @@ export default function contact() {
         </div>
 
         <div>
-          <form
-            onSubmit={sendEmail}
-            className="flex flex-col justify-start items-start"
-          >
-            <label className="font-semibold text-sm ">
-              Name<span>*</span>
-            </label>
-            <input
-              className="bg-main-green w-96 h-10 rounded mb-4 text-white px-3"
-              type="text"
-              name="from_name"
-              required
-            />
-            <label className="font-semibold text-sm ">
-              Email<span>*</span>
-            </label>
-            <input
-              className="bg-main-green w-96 h-10 rounded mb-4 text-white px-3"
-              type="email"
-              name="email"
-              required
-            />
-            <label className="font-semibold text-sm ">
-              Subject<span>*</span>
-            </label>
-            <input
-              className="bg-main-green w-96 h-10 rounded mb-4 text-white px-3"
-              type="text"
-              name="subject"
-              required
-            />
-            <label className="font-semibold text-sm ">
-              Message<span>*</span>
-            </label>
-            <textarea
-              className="bg-main-green w-96 h-30 rounded mb-4 text-white px-3"
-              name="message"
-              type="text"
-              required
-              rows="4"
-              cols="50"
-            />
-            <input
-              className="py-3 px-6 bg-main-green hover:bg-main-green-shade rounded text-white  uppercase"
-              type="submit"
-              value="Send!"
-            />
-          </form>
+          <ContactForm />
         </div>
       </div>
     </>
